backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["AlexeyTarasov77 <alexxey2707@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
gunicorn = "^22.0.0"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
requests-cache = "^1.2.1"
pandas = "^2.2.2"
openmeteo-requests = "^1.2.0"
retry-requests = "^2.0.0"
django-cors-headers = "^4.4.0"
geopy = "^2.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
